Refactor the code into modules
